Fix style of PopupModify
