+++ src/styles/theme.ts
@@ -43,7 +43,7 @@ export const theme = {
   font16: { fontSize: 16, letterSpacing: -0.4 },
   font16_bold: { fontSize: 16, letterSpacing: -0.4, fontWeight: 700 },
   font18: { fontSize: 18, letterSpacing: -0.6, fontWeight: 700 },
-  font20: { fontSize: 20, letterSpacing: -0.75 },
+  font20: { fontSize: 20, letterSpacing: -0.75, fontWeight: 700 },
 
   // Shadow
